+++ src/client/rewards/Proposition/PropositionList/PropositionListFromCatalog.js
@@ -1,11 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { get, map } from 'lodash';
+import { get, isEmpty, map } from 'lodash';
 import PropositionMainObjectCard from '../PropositionMainObjectCard';
 import CatalogBreadcrumb from '../../../object/Catalog/CatalogBreadcrumb/CatalogBreadcrumb';
 import CatalogSorting from '../../../object/Catalog/CatalogSorting/CatalogSorting';
 import Proposition from '../Proposition';
 import Loading from '../../../components/Icon/Loading';
+import OBJ_TYPE from '../../../object/const/objectTypes';
+import { parseWobjectField } from '../../../helpers/wObjectHelper';
+import { statusNoVisibleItem } from '../../../../common/constants/listOfFields';
+import CategoryItemView from '../../../object/Catalog/CategoryItemView/CategoryItemView';
+import ObjectCardView from '../../../objectCard/ObjectCardView';
 
 const PropositionListFromCatalog = ({
   intl,
@@ -25,72 +30,102 @@ const PropositionListFromCatalog = ({
   isAssign,
   user,
   history,
-  catalogGetMenuList,
   catalogHandleSortChange,
   catalogSort,
   isGetNested,
-}) => (
-  <React.Fragment>
-    {wobject && (
-      <PropositionMainObjectCard
-        intl={intl}
-        wobject={wobject}
-        currentProposition={currentProposition}
-        goToProducts={goToProducts}
-        maxReward={maxReward}
-        minReward={minReward}
-        rewardPrise={rewardPrise}
-        rewardMax={rewardMax}
-      />
-    )}
-    {isGetNested ? (
-      <Loading />
-    ) : (
-      <React.Fragment>
-        <div className="CatalogWrap__breadcrumb">
-          <CatalogBreadcrumb intl={intl} wobject={wobject} />
+  listItems,
+}) => {
+  const getListRow = listItem => {
+    const isList = listItem.object_type === OBJ_TYPE.LIST || listItem.type === OBJ_TYPE.LIST;
+    const status = get(parseWobjectField(listItem, 'status'), 'title');
+
+    if (statusNoVisibleItem.includes(status)) return null;
+    let item;
+    if (isList) {
+      item = <CategoryItemView wObject={listItem} location={location} />;
+    } else {
+      item = <ObjectCardView wObject={listItem} inList />;
+    }
+    return <div key={`category-${listItem.author_permlink}`}>{item}</div>;
+  };
+
+  const getMenuList = () => {
+    if (isEmpty(listItems)) {
+      return (
+        <div>
+          {intl.formatMessage({
+            id: 'emptyList',
+            defaultMessage: 'This list is empty',
+          })}
         </div>
-        <div className="CatalogWrap__sort">
-          <CatalogSorting
-            sort={catalogSort}
-            currWobject={wobject}
-            handleSortChange={catalogHandleSortChange}
-          />
+      );
+    }
+    return map(listItems, listItem => getListRow(listItem));
+  };
+
+  return (
+    <React.Fragment>
+      {wobject && (
+        <PropositionMainObjectCard
+          intl={intl}
+          wobject={wobject}
+          currentProposition={currentProposition}
+          goToProducts={goToProducts}
+          maxReward={maxReward}
+          minReward={minReward}
+          rewardPrise={rewardPrise}
+          rewardMax={rewardMax}
+        />
+      )}
+      {isGetNested ? (
+        <Loading />
+      ) : (
+        <React.Fragment>
+          <div className="CatalogWrap__breadcrumb">
+            <CatalogBreadcrumb intl={intl} wobject={wobject} />
+          </div>
+          <div className="CatalogWrap__sort">
+            <CatalogSorting
+              sort={catalogSort}
+              currWobject={wobject}
+              handleSortChange={catalogHandleSortChange}
+            />
+          </div>
+        </React.Fragment>
+      )}
+      <React.Fragment>
+        {map(allPropositions, propos =>
+          map(
+            propos.objects,
+            wobj =>
+              (get(wobj, ['object', 'author_permlink']) === match.params.name ||
+                get(wobj, ['object', 'parent', 'author_permlink']) === match.params.name) && (
+                <Proposition
+                  proposition={propos}
+                  wobj={wobj.object}
+                  wobjPrice={wobj.reward}
+                  assignCommentPermlink={wobj.permlink}
+                  assignProposition={assignPropositionHandler}
+                  discardProposition={discardProposition}
+                  authorizedUserName={userName}
+                  loading={loadingAssignDiscard}
+                  key={`${wobj.object.author_permlink}`}
+                  assigned={wobj.assigned}
+                  history={history}
+                  isAssign={isAssign}
+                  match={match}
+                  user={user}
+                />
+              ),
+          ),
+        )}
+        <div className="CatalogWrap">
+          <div>{getMenuList()}</div>
         </div>
       </React.Fragment>
-    )}
-    <React.Fragment>
-      {map(allPropositions, propos =>
-        map(
-          propos.objects,
-          wobj =>
-            (get(wobj, ['object', 'author_permlink']) === match.params.name ||
-              get(wobj, ['object', 'parent', 'author_permlink']) === match.params.name) && (
-              <Proposition
-                proposition={propos}
-                wobj={wobj.object}
-                wobjPrice={wobj.reward}
-                assignCommentPermlink={wobj.permlink}
-                assignProposition={assignPropositionHandler}
-                discardProposition={discardProposition}
-                authorizedUserName={userName}
-                loading={loadingAssignDiscard}
-                key={`${wobj.object.author_permlink}`}
-                assigned={wobj.assigned}
-                history={history}
-                isAssign={isAssign}
-                match={match}
-                user={user}
-              />
-            ),
-        ),
-      )}
-      <div className="CatalogWrap">
-        <div>{catalogGetMenuList()}</div>
-      </div>
     </React.Fragment>
-  </React.Fragment>
-);
+  );
+};
 
 PropositionListFromCatalog.propTypes = {
   intl: PropTypes.shape().isRequired,
@@ -110,10 +145,10 @@ PropositionListFromCatalog.propTypes = {
   isAssign: PropTypes.bool,
   user: PropTypes.shape(),
   history: PropTypes.shape().isRequired,
-  catalogGetMenuList: PropTypes.func,
   catalogHandleSortChange: PropTypes.func,
   catalogSort: PropTypes.string,
   isGetNested: PropTypes.bool,
+  listItems: PropTypes.shape(),
 };
 
 PropositionListFromCatalog.defaultProps = {
@@ -130,10 +165,10 @@ PropositionListFromCatalog.defaultProps = {
   loadingAssignDiscard: false,
   isAssign: false,
   user: {},
-  catalogGetMenuList: () => {},
   catalogHandleSortChange: () => {},
   catalogSort: '',
   isGetNested: false,
+  listItems: [],
 };
 
 export default PropositionListFromCatalog;
